Переделал до Lessons2_task3

+++ Lessons2_task1/Round.cs
@@ -12,21 +12,34 @@ namespace Lessons2_task1
         public double valueY;
         public double radius;
 
-        // Конструктор для создания объекта класса Round
-        public Round(double x, double y, double r)
+        /// <summary>
+        /// Конструктор для создания объекта класса Round
+        /// </summary>
+        /// <param name="x"></param>
+        /// <param name="y"></param>
+        /// <param name="r"></param>
+        public Round(int x, int y, int r)
         {
             SetCenter(x, y);
             SetRadius(r);
         }
 
-        // Метод для установки координат центра
+        /// <summary>
+        /// Метод для установки координат центра
+        /// </summary>
+        /// <param name="x"></param>
+        /// <param name="y"></param>
         public void SetCenter(double x, double y)
         {
             valueX = x;
             valueY = y;
         }
 
-        // Метод для установки радиуса с проверкой на корректность
+        /// <summary>
+        /// Метод для установки радиуса с проверкой на корректность
+        /// </summary>
+        /// <param name="r"></param>
+        /// <exception cref="ArgumentException"></exception>
         public void SetRadius(double r)
         {
             if (r >= 0)
@@ -39,13 +52,19 @@ namespace Lessons2_task1
             }
         }
 
-            // Метод для вычисления длины описанной окружности
+        /// <summary>
+        /// Метод для вычисления длины описанной окружности
+        /// </summary>
+        /// <returns></returns>
         public double CalculateCircumference()
         {
             return 2 * Math.PI * radius;
         }
 
-        // Метод для вычисления площади круга
+        /// <summary>
+        /// Метод для вычисления площади круга
+        /// </summary>
+        /// <returns></returns>
         public double CalculateArea()
         {
             return Math.PI * radius * radius;
